🚚 Move token list component into admin app
