+++ src/admin/views/TokensListView.js
@@ -13,8 +13,8 @@ import {
 } from 'semantic-ui-react';
 import {GET_DEV_TOKENS} from '../../state/queries';
 import {CREATE_DEV_TOKEN, DELETE_DEV_TOKEN} from '../../state/mutations';
-import {TokenList} from '../../components/TokenList';
 import NewTokenForm from '../../forms/NewTokenForm';
+import {TokenList} from '../components/TokenList';
 
 const TokensListView = () => {
   const {data, loading: devTokensLoading, error: devTokensError} = useQuery(
